move touch actions
