+++ lib/appium_lib_core/common/base.rb
@@ -13,6 +13,7 @@ require_relative 'device/screen_record'
 require_relative 'device/battery_status'
 require_relative 'device/clipboard_content_type'
 require_relative 'device/device'
+require_relative 'device/touch_actions'
 
 # The following files have selenium-webdriver related stuff.
 require_relative 'base/driver'
