+++ lib/appium_lib_core/common/base/bridge/mjsonwp.rb
@@ -16,6 +16,7 @@ module Appium
           include Device::AppState
           include Device::ScreenRecord::Command
           include Device::Device
+          include Device::TouchActions
 
           def commands(command)
             ::Appium::Core::Commands::MJSONWP::COMMANDS[command]
